using nginx to serve static files, fixes #79
