+++ nginx/Dockerfile
@@ -1,6 +1,8 @@
 FROM nginx:alpine
 
 RUN rm -rf /usr/share/nginx/html/*
+
+COPY static/ /var/www/data/static/
 COPY ./nginx.conf /etc/nginx/
 
 # Set timezone
