+++ nginx/nginx-local.conf
@@ -11,9 +11,10 @@ events { worker_connections 1024; }
 http {
 
     server {
-       listen 80;
-       listen [::]:80;
-
+        listen 80;
+        listen [::]:80;
+        
+        root /var/www/data;
         index index.php index.html index.htm;
 
         location ~ /\.ht {
@@ -30,6 +31,18 @@ http {
             proxy_set_header X-Forwarded-Proto $scheme;
             proxy_set_header Host "127.0.0.1:5565";
             proxy_pass http://masz_backend:80;
+        }
+
+        location /static {           
+            try_files $uri =404;
+        }
+
+        location /legal.txt {
+            rewrite ^ /static/legal.txt break;
+        }
+
+        location /favicon.ico {
+            rewrite ^ /static/favicon.ico break;
         }
 
         location /api {
