+++ nginx/nginx-prod.conf
@@ -14,6 +14,7 @@ http {
         listen 80;
         listen [::]:80;
 
+        root /var/www/data;
         index index.php index.html index.htm;
 
         location ~ /\.ht {
@@ -30,6 +31,18 @@ http {
             proxy_set_header X-Forwarded-Proto $scheme;
             proxy_set_header Host $http_X_Forwarded_Host;
             proxy_pass http://masz_backend:80;
+        }
+
+        location /static {
+            try_files $uri =404;
+        }
+
+        location /legal.txt {
+            rewrite ^ /static/legal.txt break;
+        }
+
+        location /favicon.ico {
+            rewrite ^ /static/favicon.ico break;
         }
 
         location /api {
