+++ nginx/static/legal.txt
