+++ webinterface/app/templates/footer.html.twig
@@ -14,7 +14,7 @@
 
         <!-- Copyright -->
         <div>
-            <a href="/api/v1/legal" target="_blank">Legal and Privacy</a>
+            <a href="/legal.txt" target="_blank">Legal and Privacy</a>
         </div>
     </div>
 </footer>
